requirements.txt
# ────────── core ──────────
python-telegram-bot[asyncio]==20.9   # бот‑API Telegram
pysnmp==4.4.12                       # SNMP‑библиотека
pyasn1==0.4.8                        # зависимость pysnmp

# ────────── опционально ──────────
# Если снова понадобится SSH‑туннель к доверенному хосту,
# раскомментируйте две строки ниже:
# sshtunnel==0.4.0                  # UDP/TCP‑форвардинг через SSH
# paramiko==3.4.0                   # SSH‑клиент, требуемый sshtunnel
